+++ changelog.md
@@ -34,6 +34,8 @@
 
 # Unreleased
 
+# 1.0.2.6
+
 - Add BlockPlaceEvent for Anti Dupe
 - Add placeholder %zmenu_test% for testing
 - Fix sound action, pitch and volume was reversed